update user repo
