+++ presentation/controller/initHandler.go
@@ -7,18 +7,17 @@ import (
 	"github.com/hrm1810884/works-hai-backend/application/usecase"
 	"github.com/hrm1810884/works-hai-backend/application/usecase/service"
 	"github.com/hrm1810884/works-hai-backend/domain"
-	"github.com/hrm1810884/works-hai-backend/infrastructure/database"
-	"github.com/hrm1810884/works-hai-backend/infrastructure/storage"
+	impl_repository "github.com/hrm1810884/works-hai-backend/infrastructure/repository"
 	"github.com/hrm1810884/works-hai-backend/ogen"
 )
 
 func (h *HaiHandler) InitGet(ctx context.Context) (ogen.InitGetRes, error) {
-	userRepository, err := database.NewImplUserRepository(ctx)
+	userRepository, err := impl_repository.NewImplUserRepository(ctx)
 	if err != nil {
 		return &ogen.InitGetBadRequest{Error: ogen.NewOptString("failed to get user repository")}, err
 	}
 
-	drawingRepository, err := storage.NewImplDrawingRepository(ctx)
+	drawingRepository, err := impl_repository.NewImplDrawingRepository(ctx)
 	if err != nil {
 		return &ogen.InitGetBadRequest{Error: ogen.NewOptString("failed to get drawing repository")}, err
 	}
